refactor(glm5): split generated text protocol
Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
